+++ samples/SampleLib/Library1.fs
@@ -4,6 +4,11 @@ open System
 
 type Text = { Value : string }
 
+type SomeContent =
+    | T of Text
+    | B of bool
+    | I of int
+
 type DataPoint =
     {
         time : DateTime